Update requirements.txt to workaround bug
There's a bug related to the cffi library. Since this is just for a
demo, I'm working around it by not using hashes in the requirements
file (as that's the source of the problem.)

In a production environment, this wouldn't be a good solution, but it's
fine for this demo.